% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemerGUI.R
\name{ggthemerGUI}
\alias{ggthemerGUI}
\title{Create ggthemer GUI}
\usage{
ggthemerGUI(plotObject = NULL)
}
\arguments{
\item{plotObject}{ggplot object, optional}
}
\value{
theme or ggplot object
}
\description{
Style your ggplots using a Shiny app
}
\examples{
ggObject <- ggplot(iris) + geom_bar(aes(x = Petal.Length, fill = Species)) # define plot aesthetitcs
myTheme <- ggthemerGUI() # Shiny module returns user theme
ggObject <- ggthemerGUI(ggObject) # View your plot within the Shiny app
ggObject <- ggObject + ggthemerGUI() # Shiny module returns user theme (data-heavy plot)

}
