% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemerSERV.R
\name{ggthemerSERV}
\alias{ggthemerSERV}
\title{Server processor for ggthemer}
\usage{
ggthemerSERV(id, theme_store)
}
\arguments{
\item{id}{Character variable}

\item{theme_store}{list}
}
\value{
reactive list
}
\description{
Server processor for ggthemer
}
\examples{

theme_store <- reactiveValues(
    myID = list(
        preview_theme = list(
            base = "grey", base_params = list(), custom_params = list()
        )
    )
)
ggthemerSERV('myID', theme_store)

}
