% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemerUI.R
\name{ggthemerUI}
\alias{ggthemerUI}
\title{Create ggthemer UI component}
\usage{
ggthemerUI(id, cached_fonts = NULL, colour_palette = NULL)
}
\arguments{
\item{id}{Character variable}

\item{cached_fonts}{list}

\item{colour_palette}{character vector}
}
\value{
shiny UI
}
\description{
Create user interface for the ggplot theme menu.
}
\examples{

cached_fonts <- list(
     Roboto = list(
         family = "Roboto",
         regular = "fonts/Roboto-Regular.ttf",
         bold = "fonts/Roboto-Bold.ttf",
         italic = "fonts/Roboto-Italic.ttf",
         bolditalic = "fonts/Roboto-BoldItalic.ttf"
    )
)
ggthemerUI('myID', cached_fonts)

}
